pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stigcode"
version = "0.1.0"
description = "SARIF-to-compliance bridge. Transforms SAST scan results into DISA STIG checklists, ATO evidence artifacts, and compliance-native output."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.12.0",
    "pyyaml>=6.0",
    "defusedxml>=0.7.0",
    "fpdf2>=2.7.0",
]
authors = [
    {name = "rdwj"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
keywords = ["stig", "disa", "compliance", "sarif", "sast", "ato", "nist", "security", "ckl"]

[project.urls]
Homepage = "https://github.com/rdwj/stigcode"
Repository = "https://github.com/rdwj/stigcode"
Issues = "https://github.com/rdwj/stigcode/issues"

[project.scripts]
stigcode = "stigcode.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"stigcode.data" = ["**/*.yaml", "**/*.xml"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "B",     # flake8-bugbear
    "UP",    # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # typer.Option/Argument in defaults is the standard Typer pattern
]
